// src/presentation/redux/actions/leadDetail/email/types.ts
export type emailResponseProps = {
  body: string;
  bodyText: string;
  cc: any;
  createTime: string;
  createdBy: string;
  deleteTime: string | null;
  emailAddress: string;
  emailIndex: number;
  name: string;
  parentId: string;
  read: boolean;
  responseTimes: number;
  subject: string;
  type: string;
  updateTime: string | null;
};